move BooleanOrString to schema.py
